+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 - Changed all APIs to use Emacs's 1-based byte positions instead of 0-based byte offsets.
+- Changed all APIs to use Emacs's 1-based line numbers instead of 0-based row numbering.
 
 ## [0.2.0] - 2020-02-02
 - Upgraded `tree-sitter` to 0.6.0.